requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.0.1
pyotp==2.9.0
qrcode==8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
openai==1.12.0
reportlab==4.0.9
Flask-Caching==2.3.1
flask-mail==0.10.0
Pillow==10.3.0
cachelib==0.13.0
Flask-Limiter==3.5.0
email-validator==2.1.0
openpyxl==3.1.5
Flasgger==0.9.7.1
APScheduler==3.10.4
psutil==5.9.8
Flask-Login==0.6.3
bcrypt==4.1.2
requests==2.31.0
python-decouple==3.8

# IA opcional
google-generativeai==0.3.2
anthropic==0.21.3
